tests/requirements.txt
Appium-Python-Client==3.1.0
selenium==4.16.0
pytest==7.4.4
